+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 
 Removed:
  - [TT-2967] Remove deprecated `Yamload::Loader#loaded_hash`
-Changed:
+ - [TT-2967] Remove schema validation
 
+Changed:
  - [TT-1790] Update ClassyHash to version 0.2.0
 
 ## 0.2.0 (2015-02-20)